requirements.txt
lmdb==1.3.0
lpips==0.1.4
numpy==1.22.3
Pillow==9.4.0
pytorch_wavelets==1.3.0
# torch==1.12.1
# torchvision==0.13.1
tqdm==4.64.0
pandas
opencv-python
PyWavelets
submitit
wandb